feat(user): add UserAddress model and update schema
Introduce the UserAddress model to store user address information, including fields for firstName, lastName, address, postalCode, phone, and country. Update the User and Country models to include relations to UserAddress. Modify the address2 field in UserAddress to be optional to accommodate cases where a secondary address line is not required.